Fix AltMode resetting
Fixes resetting Altmode combobox to `Relative` mode after loading saved plan instead of loading last entry value.

+++ GCSViews/FlightPlanner.cs
@@ -5362,7 +5362,7 @@ namespace MissionPlanner.GCSViews
                     // not home
                     if (i != 0)
                     {
-                        CMB_altmode.SelectedValue = temp.frame;
+                        CMB_altmode.SelectedValue = (int) temp.frame;
                     }
                 }
